Move enhancepcm16 into audio util
